requirements_old.txt
APScheduler==3.6.3
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
cycler==0.10.0
decorator==4.4.2
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gsheets==0.5.1
httplib2==0.18.1
idna==2.10
numpy==1.19.2
kiwisolver==1.3.1
oauth2client==4.1.3
oauthlib==3.1.0
Pillow==8.0.1
pip-autoremove==0.9.1
protobuf==3.13.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pygsheets==2.0.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==12.8
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tabulate==0.8.7
tornado==6.0.4
tzlocal==2.1
uritemplate==3.0.1
urlfetch==1.2.2
urllib3==1.25.11